requirements.txt
sqlalchemy
pydriller
pandas
gitpython